Printing the various sizes on my machine

+++ 0x00-hello_world/6-size.c
@@ -3,7 +3,7 @@
 /**
  * main - void enables the code to run successfully
  (*
- * Description: prints sizes of various types of programs on machine)?
+ * Description: prints sizes of various programs on machine)?
  * Return: 0
  */
 int main(void)
@@ -14,11 +14,11 @@ int main(void)
 	long long alonglong;
 	float afloat;
 
-	printf("Size of a char: %zu byte(s)\n" , sizeof(acharacter));
-	printf("Size of an int: %zu byte(s)\n" , sizeof(ainteger));
-	printf("Size of a long int: %zu byte(s)\n" , sizeof(along));
-	printf("Size of a long long int: %zu byte(s)" , sizeof(alonglong));
-	printf("Size of a float: %zu byte(s)" , sizeof(afloat));
+	printf("Size of a char: %zu byte(s)\n" , (unsigned int)sizeof(acharacter));
+	printf("Size of an int: %zu byte(s)\n" , (unsigned int)sizeof(ainteger));
+	printf("Size of a long int: %zu byte(s)\n" , (unsigned long)sizeof(along));
+	printf("Size of a long long int: %zu byte(s)" , (unsigned long)sizeof(alonglong));
+	printf("Size of a float: %zu byte(s)" , (unsigned int)sizeof(afloat));
 	return (0);
 }
